Edit Post Component
Now the component independently requests information about the post, it only needs an ID.

I divided the element itself into smaller components to make it easier to understand how the component works.

I also added a button to edit a post and the ability to search for posts by location, but these features are not implemented yet.